# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "appGetHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Hello world, served from the cache",
        "tags": [
          "app"
        ]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "healthLive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/LivenessDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Is the process alive? Wire restart policies here",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "healthReady",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReadinessDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Can it serve traffic? Wire healthchecks here",
        "tags": [
          "health"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "usersCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Create a user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "usersFindAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "List all users, ordered by id",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "usersFindOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Get a single user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "usersUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "usersRemove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "default": {
            "description": "Any failure. `data` is null and `error` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "Every response is an envelope: `{ data }`, or `{ data: null, error }`.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "LivenessDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ],
            "example": "ok"
          }
        },
        "required": [
          "status"
        ]
      },
      "HealthChecksDto": {
        "type": "object",
        "properties": {
          "database": {
            "enum": [
              "up",
              "down"
            ],
            "type": "string",
            "example": "up"
          },
          "cache": {
            "enum": [
              "up",
              "down"
            ],
            "type": "string",
            "example": "up"
          }
        },
        "required": [
          "database",
          "cache"
        ]
      },
      "ReadinessDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "ok",
              "degraded"
            ],
            "type": "string",
            "description": "\"degraded\" means the cache is unreachable; the api still serves requests",
            "example": "ok"
          },
          "checks": {
            "$ref": "#/components/schemas/HealthChecksDto"
          }
        },
        "required": [
          "status",
          "checks"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "description": "Trimmed and lowercased before it is stored.",
            "example": "ada@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "description": "Trimmed and lowercased before it is stored.",
            "example": "ada@example.com"
          }
        }
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "User 42 not found"
          },
          "details": {
            "description": "One entry per failed constraint, on a validation error.",
            "example": [
              "email must be an email"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-04T10:15:30.000Z"
          },
          "path": {
            "type": "string",
            "description": "The request path that failed.",
            "example": "/users/42"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "description": "Always null on the error path."
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "data",
          "error"
        ]
      }
    }
  }
}